+++ src/pages/WeeklyTargetSchedulesPage/index.vue
